+++ api/onnx_web/chain/correct_gfpgan.py
@@ -5,8 +5,8 @@ import numpy as np
 from PIL import Image
 
 from ..params import DeviceParams, ImageParams, StageParams, UpscaleParams
-from ..server.device_pool import JobContext
-from ..utils import ServerContext, run_gc
+from ..server import JobContext, ServerContext
+from ..utils import run_gc
 
 logger = getLogger(__name__)
