+++ api/onnx_web/chain/reduce_crop.py
@@ -3,8 +3,7 @@ from logging import getLogger
 from PIL import Image
 
 from ..params import ImageParams, Size, StageParams
-from ..server.device_pool import JobContext
-from ..utils import ServerContext
+from ..server import JobContext, ServerContext
 
 logger = getLogger(__name__)
