+++ api/onnx_web/chain/source_txt2img.py
@@ -7,8 +7,7 @@ from PIL import Image
 
 from ..diffusion.load import get_latents_from_seed, load_pipeline
 from ..params import ImageParams, Size, StageParams
-from ..server.device_pool import JobContext, ProgressCallback
-from ..utils import ServerContext
+from ..server import JobContext, ProgressCallback, ServerContext
 
 logger = getLogger(__name__)
