+++ api/onnx_web/chain/upscale_resrgan.py
@@ -6,8 +6,8 @@ from PIL import Image
 
 from ..onnx import OnnxNet
 from ..params import DeviceParams, ImageParams, StageParams, UpscaleParams
-from ..server.device_pool import JobContext
-from ..utils import ServerContext, run_gc
+from ..server import JobContext, ServerContext
+from ..utils import run_gc
 
 logger = getLogger(__name__)
