+++ api/onnx_web/chain/upscale_stable_diffusion.py
@@ -10,8 +10,8 @@ from ..diffusion.pipeline_onnx_stable_diffusion_upscale import (
     OnnxStableDiffusionUpscalePipeline,
 )
 from ..params import DeviceParams, ImageParams, StageParams, UpscaleParams
-from ..server.device_pool import JobContext, ProgressCallback
-from ..utils import ServerContext, run_gc
+from ..server import JobContext, ProgressCallback, ServerContext
+from ..utils import run_gc
 
 logger = getLogger(__name__)
