+++ api/onnx_web/serve.py
@@ -63,10 +63,8 @@ from .params import (
     TileOrder,
     UpscaleParams,
 )
-from .server.device_pool import DevicePoolExecutor
-from .server.hacks import apply_patches
+from .server import DevicePoolExecutor, ServerContext, apply_patches
 from .utils import (
-    ServerContext,
     base_join,
     get_and_clamp_float,
     get_and_clamp_int,
