+++ api/onnx_web/server/__init__.py
@@ -1,4 +1,10 @@
-from .device_pool import DeviceParams, DevicePoolExecutor
+from .device_pool import (
+    DeviceParams,
+    DevicePoolExecutor,
+    Job,
+    JobContext,
+    ProgressCallback,
+)
 from .hacks import (
     apply_patch_basicsr,
     apply_patch_codeformer,
@@ -6,4 +12,4 @@ from .hacks import (
     apply_patches,
 )
 from .model_cache import ModelCache
-from .upscale import run_upscale_correction
+from .context import ServerContext
